wip: initial refactor of the create kafka state machine to split trial and standard flows in separate machines
