+++ src/Kafka/CreateKafkaInstance/CreateKafkaInstance.css
@@ -1,0 +1,10 @@
+/* TextContent dl customizations */
+.mas--CreateKafkaInstance__sidebar .pf-c-content dl {
+  display: block;
+}
+.mas--CreateKafkaInstance__sidebar dt {
+  font-size: var(--pf-global--FontSize--sm);
+}
+.mas--CreateKafkaInstance__creationTimeAlert {
+  margin-top: var(--pf-global--spacer--md);
+}
