+++ src/Kafka/CreateKafkaInstance/CreateKafkaInstance.test.tsx
@@ -1,12 +1,12 @@
 import { composeStories } from "@storybook/testing-react";
 import { renderDialog, waitForI18n } from "../../test-utils";
-import * as stories from "./CreateKafkaInstanceWithSizes.stories";
+import * as stories from "./CreateKafkaInstance.stories";
 
-const { CreateKafkaInstanceWithSizes } = composeStories(stories);
+const { CreateKafkaInstance } = composeStories(stories);
 
-describe("CreateKafkaInstanceWithSizes", () => {
+describe("CreateKafkaInstance", () => {
   it("should persist ouiaComponentId of create instance button", async () => {
-    const comp = renderDialog(<CreateKafkaInstanceWithSizes />);
+    const comp = renderDialog(<CreateKafkaInstance />);
     await waitForI18n(comp);
 
     const btnSubmit = comp.getByRole("button", { name: "Create instance" });
