+++ src/Kafka/CreateKafkaInstance/Stories/FormProcessing.stories.tsx
@@ -1,5 +1,5 @@
 import type { ComponentMeta } from "@storybook/react";
-import { CreateKafkaInstanceWithSizes } from "../CreateKafkaInstanceWithSizes";
+import { CreateKafkaInstance } from "../CreateKafkaInstance";
 import {
   argTypes,
   parameters,
@@ -9,7 +9,7 @@ import {
 } from "./storiesHelpers";
 
 export default {
-  component: CreateKafkaInstanceWithSizes,
+  component: CreateKafkaInstance,
   args: {
     apiPlan: "standard",
     apiScenario: "standard-available",
@@ -23,7 +23,7 @@ export default {
   },
   argTypes,
   parameters,
-} as ComponentMeta<typeof CreateKafkaInstanceWithSizes>;
+} as ComponentMeta<typeof CreateKafkaInstance>;
 
 export const LoadingData = Template.bind({});
 LoadingData.args = {
