+++ src/Kafka/CreateKafkaInstance/components/InstanceInfo.tsx
