+++ src/Kafka/CreateKafkaInstance/components/InstanceInfoSkeleton.tsx
