+++ src/Kafka/CreateKafkaInstance/index.ts
@@ -1,0 +1,2 @@
+export * from "./CreateKafkaInstance";
+export * from "./types";
