--- src/Kafka/CreateKafkaInstanceWithSizes/components/CloudProviderTiles.css
@@ -1,7 +1,0 @@
-.mas--CreateKafkaInstanceWithSizes__CloudProviderTile--Icon {
-  margin-right: var(--pf-global--spacer--md);
-}
-
-.mas--CreateKafkaInstanceWithSizes__CloudProviderTile--select {
-  display: none;
-}
