--- src/Kafka/CreateKafkaInstanceWithSizes/index.ts
@@ -1,2 +1,0 @@
-export * from "./CreateKafkaInstanceWithSizes";
-export * from "./types";
