+++ src/Kafka/index.ts
@@ -1,4 +1,4 @@
-export * from "./CreateKafkaInstanceWithSizes";
+export * from "./CreateKafkaInstance";
 export * from "./KafkaInstanceDrawer";
 export * from "./KafkaInstanceStatus";
 export * from "./KafkaMessageBrowser";
